Borders Fixed.
Concept: when you collide into the border, you are transported to x = 50, which is IN the border. Therefore, transport yourself outside the border, so when you collide again, it stops you. This also happens for the top and bottom borders as well, so they have been fixed. Very happy right now. :D